:heavy_plus_sign: Add styled-component
